move more stuff to mapobject
